pyproject.toml
[tool.poetry]
name = "discord-dicebot"
version = "0.1.0"
description = ""
authors = ["tksnnx <tkser.edu@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
dice = "^4.0.0"
python-dotenv = "^1.0.1"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
